add signin and signup without testing

+++ mobilityBackend/src/main/java/com/gurus/mobility/payload/response/JwtResponse.java
@@ -1,17 +1,18 @@
 package com.gurus.mobility.payload.response;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import java.util.List;
 
 public class JwtResponse {
 
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private Integer id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, Long id, String username, String email, List<String> roles) {
+    public JwtResponse(String accessToken, Integer id, String username, String email, List<String> roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
@@ -35,11 +36,11 @@ public class JwtResponse {
         this.type = tokenType;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
